requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg[binary]>=3.1.0
sentence-transformers>=2.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary
pgvector